@@ -5,13 +5,16 @@
   "dependencies": {
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
+    "@material-ui/lab": "^4.0.0-alpha.56",
+    "axios": "^0.19.2",
     "clsx": "^1.1.1",
     "formik": "^2.1.4",
     "formik-material-ui": "^2.0.1",
+    "jwt-decode": "^2.2.0",
     "material-ui": "^0.20.2",
     "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-image": "^4.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^2.1.8",
