@@ -20,10 +20,10 @@ const UserSchema = new Schema({
 		type: String,
 		required: true
 	},
-	emailVerifed:{
+	emailVerified:{
 		type: Boolean,
 		required: true
 	}
 });
 
-module.exports = User = mongoose.model('user', UserSchema);
+module.exports = mongoose.model('user', UserSchema);
