@@ -1,4 +1,4 @@
 module.exports = {
 	secretKey: 'jwt-secret',
-	timeout: '1 hr'
+	timeout: '1 day'
 };  
